refactor: rename @types folder
